added get_histogram_stat_spec to ValidateReleaseUtil #683
